requirements.txt
fastapi
uvicorn
groq
supabase
twilio
python-dotenv